Или так, например

+++ src/main/java/net/proselyte/springbootdemo/repository/UserRepository.java
@@ -3,5 +3,7 @@ package net.proselyte.springbootdemo.repository;
 import net.proselyte.springbootdemo.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+
+//я так понял, DAO нам теперь не нужны?)
+public interface UserRepository extends JpaRepository<User, Integer> {
 }
